+++ typescript/infra/config/environments/mainnet3/chains.ts
@@ -57,6 +57,20 @@ export const chainMetadataOverrides: ChainMap<Partial<ChainMetadata>> = {
       minPriorityFeePerGas: 1 * 10 ** 9, // 1 gwei
     },
   },
+  ronin: {
+    transactionOverrides: {
+      minGasPrice: 20 * 10 ** 9, // 20 gwei
+      minFeePerGas: 20 * 10 ** 9, // 20 gwei
+      minPriorityFeePerGas: 20 * 10 ** 9, // 20 gwei
+    },
+  },
+  ink: {
+    transactionOverrides: {
+      minGasPrice: 1, // 1 wei
+      minFeePerGas: 1, // 1 wei
+      minPriorityFeePerGas: 1, // 1 wei
+    },
+  },
   // Deploy-only overrides, set when deploying contracts
   // chilizmainnet: {
   //   transactionOverrides: {